Remove sshd from images
Originally, sshd was used as a way to run and debug services via IDEs because native docker/docker-compose support was generally bad. Moreover, the variables.sh setup prevented most IDE setups in the first place.
With the new template-based docker-compose.yml setup, sshd is not needed anymore and is in general just a security risk.